API search request

+++ src/main/java/com/minswap/hrms/entities/Evidence.java
@@ -19,7 +19,7 @@ public class Evidence {
     private int evidenceId;
 
     @Column(name = "request_id")
-    private int requestId;
+    private Long requestId;
 
     @Column(name = "image")
     private String image;
